Moved UserEdit to domain package
