+++ src/com/io/gui/EditorEvent.java
@@ -1,5 +1,7 @@
 package com.io.gui;
 
+import com.io.domain.UserEdit;
+
 public interface EditorEvent {
     void sendChange(UserEdit useredit);
 }
