+++ src/com/io/gui/StartIo.java
@@ -4,6 +4,7 @@ import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
 import com.intellij.openapi.actionSystem.LangDataKeys;
 import com.intellij.openapi.editor.Editor;
+import com.io.domain.UserEdit;
 import com.io.net.Server;
 import com.io.net.ConnectorEvent;
