+++ src/com/io/gui/StartReceiving.java
@@ -5,6 +5,7 @@ import com.intellij.openapi.editor.Document;
 import com.intellij.openapi.editor.Editor;
 import com.intellij.openapi.editor.event.DocumentListener;
 import com.intellij.openapi.project.Project;
+import com.io.domain.UserEdit;
 
 public class StartReceiving {
