+++ src/com/io/net/ConnectorEvent.java
@@ -1,6 +1,6 @@
 package com.io.net;
 
-import com.io.gui.UserEdit;
+import com.io.domain.UserEdit;
 
 public interface ConnectorEvent {
     void applyUserEdit(UserEdit userEdit);
